Add missing Libs to pkgconfig file
I'm not sure if these should be in Libs or Libs.private but it seems
they need to be in Libs or all executables need to explicitly link to
whatever libraries portaudio links to, but then how can they know how
portaudio is configured etc. So seems like they should be in Libs.

-lasound is in Libs on linux so perhaps this is correct.